Test added

+++ FinalTask/src/main/java/by/training/finaltask/service/impl/StudentServiceImpl.java
@@ -1,8 +1,6 @@
 package by.training.finaltask.service.impl;
 
-import by.training.finaltask.bean.entities.Dean;
-import by.training.finaltask.bean.entities.Document;
-import by.training.finaltask.bean.entities.Student;
+import by.training.finaltask.bean.entities.*;
 import by.training.finaltask.dao.DaoFactory;
 import by.training.finaltask.dao.daoimpl.DeanDaoImpl;
 import by.training.finaltask.dao.daoimpl.DocumentDaoImpl;
@@ -15,10 +13,7 @@ import by.training.finaltask.service.excpetion.ServiceException;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
-import java.util.Date;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 import java.util.regex.Pattern;
 
 public class StudentServiceImpl extends BaseService implements StudentService {
@@ -26,12 +21,24 @@ public class StudentServiceImpl extends BaseService implements StudentService {
     private static final Logger debugLog = LogManager.getLogger("DebugLog");
 
     @Override
-    public Student viewInfo(Integer id) throws ServiceException {
+    public Student viewInfo(Integer id, boolean isAdmin) throws ServiceException {
+        if(id == null){
+            return null;
+        }
         StudentDaoImpl studentDao = DaoFactory.getInstance().getStudentDao();
-        transaction.init(studentDao);
+        UserDaoImpl userDao = DaoFactory.getInstance().getUserDao();
+        transaction.init(studentDao, userDao);
         debugLog.debug("in find");
         try {
-            return studentDao.findById(id);
+            Student student = studentDao.findById(id);
+            User user = userDao.findById(id);
+            if(user.getRole() == Role.STUDENT) {
+                return student;
+            }
+            if(isAdmin && user.getRole() != Role.ADMINISTRATOR){
+                return student;
+            }
+            return null;
         } catch (DaoException e) {
             throw new ServiceException(e);
         } finally {
@@ -42,10 +49,24 @@ public class StudentServiceImpl extends BaseService implements StudentService {
     @Override
     public boolean updateInfo(Student student) throws ServiceException {
         if (student != null) {
+            if(!studentCheck(student)){
+                debugLog.debug("student dont match");
+                return false;
+            }
             StudentDaoImpl studentDao = DaoFactory.getInstance().getStudentDao();
-            transaction.init(studentDao);
+            UserDaoImpl userDao = DaoFactory.getInstance().getUserDao();
+            transaction.init(studentDao, userDao);
             try {
-                return studentDao.update(student);
+                if(student.getId() != null){
+                    User user = userDao.findById(student.getId());
+                    if(user.getRole() == Role.ADMINISTRATOR){
+                        debugLog.debug("student is admin");
+                        return false;
+                    }
+                    return studentDao.update(student);
+                }
+                debugLog.debug("id = null");
+               return false;
             } catch (DaoException e) {
                 transaction.rollback();
                 throw new ServiceException(e);
@@ -62,6 +83,9 @@ public class StudentServiceImpl extends BaseService implements StudentService {
     public List<Student> viewDeanStudents(Integer id) throws ServiceException {
         StudentDaoImpl studentDao = DaoFactory.getInstance().getStudentDao();
         transaction.init(studentDao);
+        if(id == null){
+            return null;
+        }
         try {
             return studentDao.findDeanStudent(id);
         } catch (DaoException e) {
@@ -73,6 +97,7 @@ public class StudentServiceImpl extends BaseService implements StudentService {
 
     @Override
     public Map<Student, Dean> viewStudentsInfo() throws ServiceException {
+        debugLog.debug("in service");
         try {
             Map<Student, Dean> studentDeanMap = new HashMap<>();
             StudentDaoImpl studentDao = DaoFactory.getInstance().getStudentDao();
@@ -127,17 +152,20 @@ public class StudentServiceImpl extends BaseService implements StudentService {
         if (id != null && id > 0) {
             StudentDaoImpl studentDao = DaoFactory.getInstance().getStudentDao();
             DocumentDaoImpl documentDao = DaoFactory.getInstance().getDocumentDao();
-            transaction.init(studentDao, documentDao);
+            UserDaoImpl userDao = DaoFactory.getInstance().getUserDao();
+            transaction.init(studentDao, documentDao, userDao);
             try {
+                User user = userDao.findById(id);
+                if(user == null || user.getRole() == Role.ADMINISTRATOR){
+                    return false;
+                }
+
                 if (studentDao.findById(id) != null) {
                     List<Document> documents = documentDao.findByUserId(id);
                     if (documents.size() > 0) {
-                        for (Document document : documents) {
                             if (!documentDao.deleteUserReferences(id)) {
-                                debugLog.debug("cannot update doc " + document);
                                 return false;
                             }
-                        }
                     }
                     return studentDao.delete(id);
                 } else {
@@ -157,16 +185,20 @@ public class StudentServiceImpl extends BaseService implements StudentService {
     private boolean studentCheck(Student student) {
         Pattern pattern = Pattern.compile("^[a-zA-Z0-9_+&*-]+(?:.[a-zA-Z0-9_+&*-]" +
                 "+)*@(?:[a-zA-Z0-9-]+\\.)+[a-zA-Z]{2,7}");
-        if (!pattern.matcher(student.getMail()).matches()) {
+        boolean result = pattern.matcher(student.getMail()).matches();
+        if (!result) {
+            debugLog.debug("mail dont macth" + student.getMail());
             return false;
         }
         if (student.getName() == null || student.getDeanId() == null || student.getDate() == null ||
                 student.getLastname() == null || student.getPatronymic() == null) {
+            debugLog.debug("something dont match");
             return false;
         }
 
         pattern = Pattern.compile("\\d{4}-\\d{2}-\\d{2}");
         if (!pattern.matcher(student.getDate()).matches()) {
+            debugLog.debug("data dont macth");
             return false;
         }
         return true;
